#169
- common/modifier: реструктуризация раздела
- мелкие правки стиля кода
