+++ src/common/components/modifier/index.ts
@@ -1,0 +1,9 @@
+export type {
+  TextContent,
+  ImageContent,
+  ColorContent,
+  ModifierContent,
+  ModifierProps,
+  MoreButtonProps,
+} from './types';
+export { Modifier, MoreButton } from './modifier';
