+++ src/common/components/product-info/types.ts
@@ -1,7 +1,9 @@
+import type { ReactNode } from 'react';
+
 export interface ProductInfoProps {
   /** Ограничение товара. Влияет на отображение информации. */
   restriction?: 'adult' | 'unavailable';
 
   /** Содержимое. */
-  children?: React.ReactNode;
+  children?: ReactNode;
 }
