+++ src/mobile/components/select-screen/index.tsx
@@ -1,4 +1,4 @@
-import React, { Children, isValidElement } from 'react';
+import React, { ButtonHTMLAttributes, Children, isValidElement } from 'react';
 import { LayoutProps, MobileLayout } from '@sima-land/ui-nucleons/layout';
 import CheckSVG from '@sima-land/ui-quarks/icons/24x24/Stroked/Check';
 import cn from 'classnames';
@@ -6,7 +6,7 @@ import styles from './select-screen.module.scss';
 
 export type OptionSize = 's' | 'm' | 'l' | 'xl';
 
-export interface SelectScreenOptionProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
+export interface SelectScreenOptionProps extends ButtonHTMLAttributes<HTMLButtonElement> {
   size?: OptionSize;
   selected?: boolean;
   withSelectedIcon?: boolean;
@@ -17,34 +17,38 @@ export interface SelectScreenOptionProps extends React.ButtonHTMLAttributes<HTML
  * @param props Свойства.
  * @return Элемент.
  */
-export const SelectScreenLayout = ({ children, ...props }: LayoutProps) => (
-  <MobileLayout {...props} disabledOn={['mxs', 'ms']}>
-    <div className={styles.container}>
-      {Children.toArray(children).filter(x => isValidElement(x) && x.type === SelectScreenOption)}
-    </div>
-  </MobileLayout>
-);
+export function SelectScreenLayout({ children, ...props }: LayoutProps) {
+  return (
+    <MobileLayout {...props} disabledOn={['mxs', 'ms']}>
+      <div className={styles.container}>
+        {Children.toArray(children).filter(x => isValidElement(x) && x.type === SelectScreenOption)}
+      </div>
+    </MobileLayout>
+  );
+}
 
 /**
  * Опция.
  * @param props Свойства.
  * @return Элемент.
  */
-export const SelectScreenOption = ({
+export function SelectScreenOption({
   children,
   className,
   selected,
   withSelectedIcon = selected,
   size = 'm',
   ...props
-}: SelectScreenOptionProps) => (
-  <button
-    {...props}
-    type='button'
-    className={cn(styles.item, className, styles[`size-${size}`])}
-    data-testid='select-screen:option'
-  >
-    {children}
-    {selected && withSelectedIcon && <CheckSVG role='presentation' className={styles.icon} />}
-  </button>
-);
+}: SelectScreenOptionProps) {
+  return (
+    <button
+      {...props}
+      type='button'
+      className={cn(styles.item, className, styles[`size-${size}`])}
+      data-testid='select-screen:option'
+    >
+      {children}
+      {selected && withSelectedIcon && <CheckSVG role='presentation' className={styles.icon} />}
+    </button>
+  );
+}
